requirements.txt
chainlit>=2.8.0
openai>=1.54.3
chromadb>=0.5.15
python-dotenv>=1.0.1
PyPDF2>=3.0.1
tavily-python>=0.5.0
